Fix log.cpp
git-svn-id: svn://svn.code.sf.net/p/boomerang/code/trunk@2441 dc5e1367-85d4-4bd1-96b4-a20bfbfd92d5

+++ boomerang/log.cpp
@@ -72,7 +72,7 @@ Log &Log::operator<<(LocationSet *l)
   return *this;
 }
 
-Log &Log::operator<<(int i)
+Log &Log::operator<<(intptr_t i)
 {
   std::ostringstream st;
   st << std::dec << i;
